# perf: add caches to reduce python overhead