--- a/src/awkward/_meta/unionmeta.py
+++ b/src/awkward/_meta/unionmeta.py
@@ -58,12 +58,14 @@ class UnionMeta(Meta, Generic[T]):
     def minmax_depth(self) -> tuple[int, int]:  # type: ignore[override]
         if len(self._contents) == 0:
             return (0, 0)
-        mins, maxs = [], []
-        for content in self._contents:
-            mindepth, maxdepth = content.minmax_depth
-            mins.append(mindepth)
-            maxs.append(maxdepth)
-        return (min(mins), max(maxs))
+        mindepth, maxdepth = self._contents[0].minmax_depth
+        for content in self._contents[1:]:
+            mindepth_, maxdepth_ = content.minmax_depth
+            if mindepth_ < mindepth:
+                mindepth = mindepth_
+            if maxdepth_ > maxdepth:
+                maxdepth = maxdepth_
+        return (mindepth, maxdepth)
 
     @cached_property
     def branch_depth(self) -> tuple[bool, int]:  # type: ignore[override]