This PR adds some caches to reduce python overhead. This does not affect any "number crunching" with arrays. Thus, the main benefit is for typetracing although I noticed a slight speed difference for the cpu backend aswell. 

I tested this with a NanoAOD MC sample of 100k events and the trijet mass calculation from the AGC (ref: https://github.com/iris-hep/idap-200gbps/blob/main/agc-coffea-2024.ipynb, cell 2).

I measured the runtime, peak memory consumption, accumulated memory consumption (total memory), and number of allocations applying the trijet mass calculation to a `coffea.NanoEvents` record array and got the following:

### CPU backend
|         | Runtime | Peak memory | Total memory | # Allocations |
|---------|---------|-------------|------------|---------------|
| main    |   29.3 ms ± 434 μs      |   7.025 MB          | 37.769 MB |     3691         |
| this PR (49f99f2c7fd6bf731d90feccd7b24e420446dbf8) |   25.6 ms ± 260 μs     |     7.025 MB        |  36.659 MB |     2975        |
| this PR (dc89223f5c68e3dcb488b1e9d298088f3bbb6011) |   24.8 ms ± 167 μs     |     7.025 MB        |  36.538 MB |     2915        |
| **this PR (a6f2c8f805bbea0d036dab02384b06403608646f, final)** |  **25.1 ms ± 126 μs**     |     **7.025 MB**        | **36.390 MB** |     **2633**        |


### TypeTracer backend
|         | Runtime | Peak memory | Total memory | # Allocations |
|---------|---------|-------------|---|------------|
| main    |   30.3 ms ± 145 μs      |   17.405 KB      | 2.609 MB    |      2882         |
| this PR (49f99f2c7fd6bf731d90feccd7b24e420446dbf8) |   25.3 ms ± 136 μs    |     17.405 KB     | 1.495 MB   |       2164        |
| this PR (dc89223f5c68e3dcb488b1e9d298088f3bbb6011) |   25.2 ms ± 135 μs     |     17.405 KB        |  1.375 MB |     2104        |
| **this PR (a6f2c8f805bbea0d036dab02384b06403608646f, final)** |   **24.4 ms ± 195 μs**     |     **16.336 KB**        |  **1.224MB** |     **1816**        |

I think this is a noticeable speedup (~25% faster in this example) for typetracing, which ultimately improves the user experience with dask-awkward.