@pfackeldey 2025-01-06
Oh there seems to be something new with mypy and the typetracer. I'm pretty sure this is unrelated to this PR though...