@pfackeldey 2025-01-07
I'm done optimizing. Looking forward to your review & feedback @jpivarski. I structured the table such that you can see how much each commit brings to the table.