@pfackeldey 2025-01-07
Hi @ianna,
Total memory is the cumulative sum of all memory allocated during the process: less allocations lead to less cumulated memory consumption. The speedup comes of course from hitting caches and not doing python things. The cache does exist a certain time and has some memory, but it doesn't increase the peak memory consumption. I think this is because it doesn't matter there if it is allocated on the fly or already there in the cache. There might be times where the cache is existing, but not needed, this would increase slightly the "instantaneous" memory consumption. I'd argue this doesn't matter too much because a) the caches are pretty small and b) the peak memory is not increasing (this is the most important metric).

Concerning the cache lifetimes:

### `@cached_property`

These caches are bound to the instance and thus are deleted when the instance is deleted. They only cache a tuple of two integers, which is just a few bytes, or a list of strings (fields), also just a few bytes. This shouldn't be any problem.

### `@lru_cache`

These caches live until the python process is closed (or you actively clear them). They cache function arguments with their hash (just a python int) and the function's return value(s). Here, the return values can be slightly larger:
- `_broadcast_shapes`: tuple of ints -> O(bytes)
- `primitive_to_dtype`: numpy dtype -> O(bytes)
-  `dtype_to_primitive`: str -> O(bytes)
- `from_json`: ak.forms.Form -> can range from O(bytes) to O(kilo bytes), e.g. `coffea.NanoEvents` form (for MC) is ~580 KB (if serialized as a string), the form itself is a lot smaller - but I might not measure it correctly. This cache is usually hit in `ak.from_iter` with the ArrayBuilder. I can see rare scenarios  (i.e. _many different large_ `ak.forms.Forms`) where this cache can become a bit large (O(MB)), but I don't have a good feeling about how realistic this is. In these scenarios it might actually be very beneficial for the runtime to have this cache, because such large forms might take some time to recursively iterate over... (see: https://github.com/scikit-hep/awkward/blob/main/src/awkward/forms/form.py#L54-L202)
If you want, I can remove this one, I have no strong opinions about it. (If I recall correctly this cache did not significantly contribute to the speedups shown in the tables from above)

Concerning stale caches: I don't think it can happen that the caches are not updated correctly. The cache keys are based on the the hash of the function arguments and they are well-defined and unique (e.g. tuples of ints). So I'd argue there's nothing to worry about stale caches, unless there's a bug in `functools`, which is something we can rely on.