@pfackeldey 2025-01-07
And a side note for future reference:
I tested adding `__slots__` to all meta, index, and content classes, but I saw no measurable speedup.