@pfackeldey 2025-01-07
> ```
> In [3]: sys.getsizeof(foo)
> Out[3]: 48
> 
> In [4]: sys.getsizeof(fooslots)
> Out[4]: 40
> ```

I'll give `__slots__` another try and see if it matters for memory. :+1: